requirements.txt
matplotlib
numpy
pandas
python-dotenv
tqdm
tika